Add description, ingredients and steps
Added an image of palak paneer along with a heading to the top of the
page. Second change was adding a description heading along with a
description paragraph of the recipe. The third step was adding an
ingredients heading along with an unordered list of said ingreidents.
Finally the last change was to add a steps heading along with an ordered
list of the steps needed for making the dish.