Image, description, ingredients, and steps added
The first change was an image of aloo gobi added to the aloo-gobi.html
file. The second change was a heading description and a paragraph of
description about aloo gobi. Next was an ingredients heading, and an
unordered list of those ingredients. Lastly, a steps heading and an
ordered list of those steps was added.